Linux port hardening from the parallel session: install/deployment paths that
never worked, nine drivers whose library probes had macOS-only paths, and a set
of correctness bugs that were latent on macOS too --

  * xtime.cpp   asm("rdtsc" : "=A") is the EDX:EAX pair only on i386; on
                x86-64 timeStamp() returned the low 32 bits and wrapped every
                ~1.4 s at 3 GHz, feeding the negotiation watchdog and the RT
                WCET histograms random negative deltas.
  * xthread.cpp XCondition::wait() normalized tv_nsec with one conditional
                subtraction, so any timeout >= ~1 s returned EINVAL at once --
                a timed wait degenerating into a busy spin.
  * main.cpp    deque iterator used after push_back (UB), infinite loop when
                two search roots both held libpython.
  * support.cpp negative char passed to isprint() (UB); every non-ASCII byte in
                the interface debug log printed wrong.

Plus the embedded IPython kernel working on ipykernel 7 (the %gui hook now
returns when a shell request is pending, as every stock loop_* does), and the
first GUI verification of the Linux port on a real X server.

Merges clean; the five files both sides touched (kame/main.cpp,
kamepoolalloc/{README.md,allocator.cpp,design/RT_READINESS.md,tests/bench/bench_rt_wcet.cpp})
are disjoint edits.  kamestm/ STM work on this side is untouched by it.

+++ kame/script/xpythonmodule.cpp
@@ -360,7 +360,8 @@ KAMEPyBind::export_embedded_module_basic(pybind11::module_& m) {
         .def("__int__", [](XULongNode::Payload &self)->unsigned long{return self;});}
     {   auto [node, payload] = XPython::bind.export_xvaluenode<XHexNode, unsigned long, XValueNodeBase>("XHexNode");
         (*payload)
-        .def("__int__", [](XHexNode::Payload &self)->unsigned int{return self;});}
+        //!< `unsigned long`, not `unsigned int` — the node's value type (see XHexNode's str_).
+        .def("__int__", [](XHexNode::Payload &self)->unsigned long{return self;});}
     {   auto [node, payload] = XPython::bind.export_xvaluenode<XBoolNode, bool, XValueNodeBase>("XBoolNode");
         (*payload)
         .def("__bool__", [](XBoolNode::Payload &self)->bool{return self;});}
